rename directory to notifications
