+++ src/webapp/components/notifications/NotificationDetailsDialog.tsx
