+++ src/webapp/components/notifications/useNotificationDetailsDialog.ts
